LinuxContainer: Move Process type out of Configuration
As a first step to making a pod type, move the nested process
configuration out of LinuxContainer.Configuration.Process and
into a standalone LinuxProcessConfiguration type.
